requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
pytest
plotly
xgboost
catboost
lightgbm
umap-learn
category-encoders